+++ core/include/scipp/core/element/comparison.h
@@ -10,26 +10,45 @@
 #include "scipp/common/overloaded.h"
 #include "scipp/core/element/arg_list.h"
 #include "scipp/core/transform_common.h"
+#include "scipp/core/values_and_variances.h"
 
 /// Operators to be used with transform and transform_in_place to implement
 /// operations for Variable.
 namespace scipp::core::element {
 
-constexpr auto is_approx = overloaded{
-    transform_flags::no_out_variance,
-    arg_list<double, float, int64_t, int32_t, std::tuple<double, double, float>,
-             std::tuple<int64_t, int32_t, int32_t>,
-             std::tuple<int64_t, int64_t, int32_t>,
-             std::tuple<int64_t, int32_t, int64_t>,
-             std::tuple<int32_t, int32_t, int64_t>,
-             std::tuple<int32_t, int64_t, int64_t>>,
-    [](const units::Unit &x, const units::Unit &y, const units::Unit &t) {
-      expect::equals(x, y);
-      expect::equals(x, t);
-      return units::dimensionless;
-    },
-    [](const auto &x, const auto &y, const auto &t) {
+using isclose_types_t = arg_list_t<
+    double, float, int64_t, int32_t, std::tuple<float, float, double>,
+    std::tuple<int64_t, int64_t, double>, std::tuple<int32_t, int32_t, double>,
+    std::tuple<int32_t, int64_t, double>, std::tuple<int64_t, int32_t, double>,
+    std::tuple<int64_t, int32_t, int64_t>,
+    std::tuple<int32_t, int32_t, int64_t>,
+    std::tuple<int32_t, int64_t, int64_t>>;
+
+constexpr auto isclose_units = [](const units::Unit &x, const units::Unit &y,
+                                  const units::Unit &t) {
+  expect::equals(x, y);
+  expect::equals(x, t);
+  return units::dimensionless;
+};
+
+constexpr auto isclose = overloaded{
+    transform_flags::expect_no_variance_arg_t<2>{}, isclose_types_t{},
+    isclose_units, [](const auto &x, const auto &y, const auto &t) {
       using std::abs;
+      return abs(x - y) <= t;
+    }};
+
+constexpr auto isclose_equal_nan = overloaded{
+    transform_flags::expect_no_variance_arg_t<2>{}, isclose_types_t{},
+    isclose_units, [](const auto &x, const auto &y, const auto &t) {
+      using std::abs;
+      using numeric::isnan;
+      using numeric::isinf;
+      using numeric::signbit;
+      if (isnan(x) && isnan(y))
+        return true;
+      if (isinf(x) && isinf(y) && signbit(x) == signbit(y))
+        return true;
       return abs(x - y) <= t;
     }};