+++ core/test/element_util_test.cpp
@@ -65,17 +65,20 @@ TEST(ElementUtilTest, convertMaskedToZero_accepts_all_types) {
       std::is_same_v<decltype(convertMaskedToZero(int64_t{}, true)), int64_t>);
 }
 
-TEST(ElementUtilTest, values_variances) {
+TEST(ElementUtilTest, values_variances_stddev) {
   ValueAndVariance x{1.0, 2.0};
   EXPECT_EQ(values(units::m), units::m);
   EXPECT_EQ(values(x), 1.0);
   EXPECT_EQ(values(1.2), 1.2);
   EXPECT_EQ(variances(units::m), units::m * units::m);
   EXPECT_EQ(variances(x), 2.0);
+  EXPECT_EQ(stddevs(units::m), units::m);
+  EXPECT_EQ(stddevs(units::counts), units::counts);
+  EXPECT_EQ(stddevs(x), sqrt(2.0));
 }
 
 namespace {
-constexpr auto test_is_sorted = [](const auto sorted, const bool order) {
+constexpr auto test_issorted = [](const auto sorted, const bool order) {
   const auto expect_sorted_eq = [&sorted](const auto a, const auto b,
                                           const auto expected) {
     bool out = true;
@@ -96,32 +99,32 @@ constexpr auto test_is_sorted = [](const auto sorted, const bool order) {
 };
 }
 
-TEST(ElementUtilTest, is_sorted) {
-  test_is_sorted(is_sorted_nondescending, true);
-  test_is_sorted(is_sorted_nonascending, false);
+TEST(ElementUtilTest, issorted) {
+  test_issorted(issorted_nondescending, true);
+  test_issorted(issorted_nonascending, false);
 }
 
-TEST(ElementUtilTest, is_linspace_time_point) {
-  const auto is_linspace_wrapper = [](std::initializer_list<int64_t> ints) {
+TEST(ElementUtilTest, islinspace_time_point) {
+  const auto islinspace_wrapper = [](std::initializer_list<int64_t> ints) {
     std::vector<core::time_point> vec;
     vec.reserve(ints.size());
     for (auto x : ints) {
       vec.emplace_back(x);
     }
 
-    return core::element::is_linspace(span<const core::time_point>(vec));
+    return core::element::islinspace(span<const core::time_point>(vec));
   };
 
-  EXPECT_FALSE(is_linspace_wrapper({}));
-  EXPECT_FALSE(is_linspace_wrapper({0}));
-  EXPECT_FALSE(is_linspace_wrapper({0, 1, 1}));
-  EXPECT_FALSE(is_linspace_wrapper({0, 0, 1}));
-  EXPECT_FALSE(is_linspace_wrapper({1, 1, 1}));
-  EXPECT_FALSE(is_linspace_wrapper({0, 1, 3}));
+  EXPECT_FALSE(islinspace_wrapper({}));
+  EXPECT_FALSE(islinspace_wrapper({0}));
+  EXPECT_FALSE(islinspace_wrapper({0, 1, 1}));
+  EXPECT_FALSE(islinspace_wrapper({0, 0, 1}));
+  EXPECT_FALSE(islinspace_wrapper({1, 1, 1}));
+  EXPECT_FALSE(islinspace_wrapper({0, 1, 3}));
 
-  EXPECT_TRUE(is_linspace_wrapper({0, 1}));
-  EXPECT_TRUE(is_linspace_wrapper({0, 1, 2}));
-  EXPECT_TRUE(is_linspace_wrapper({0, 2, 4, 6}));
+  EXPECT_TRUE(islinspace_wrapper({0, 1}));
+  EXPECT_TRUE(islinspace_wrapper({0, 1, 2}));
+  EXPECT_TRUE(islinspace_wrapper({0, 2, 4, 6}));
 }
 
 TEST(ElementUtilTest, zip) {